Clear waitingEPI entry if endpoint is /unregistered/.
This is still a terrible kludge, but this fix at least
means if a localization attempt fails and then
the process is closed (perhaps during a benchmark run :P)
later attempts to kludge-pair won't try to match
the stale endpoint.

Concurrent attempts are as broken as ever still, though, of course. :)