Inter-container TCP optimization working! :D
* Run ipcd in dedicated container (because we can :P)
* Run server/client in normal ipc-able containers
  * Contains libipc.so and /etc/ld.so.preload to ensure it's used
* When starting server/client, give docker single extra flag
  * This flag bind-mounts the UDS for talking to ipcd
    into each of the participating containers so they
    all talk to the same 'central manager'.

That's it!

During optimization phase, ipcd normally creates
a connected pair of sockets and passes each
process one end of this pair for use as their
direct fast transport.  This open-fd passing
is done via UDS, so since we're able to establish
our UDS into the containers this mechanism
still works!

This is cool because it gives the containers
a means for talking directly despite there
being no existing way for them to open
connected sockets in this way themselves!