libipc: Issue shm_unlink while still open, hopefully fix perm issues.
Previously, since it seems... 'cleaner', we didn't issue shm_unlink
until our state had been successfully restored.

'Cleaner' since it seems strange to request the OS delete
a memory object while you still have intentions of using it :)

This has two problems:
1) State is never restored, so memory object lives forever.
   This may happen in the process of disabling ipcd (maybe?),
   if the program crashes, and maybe if the program executes
   to completion without ever interacting with any symbols
   that would cause libipc.so to be loaded (maybe).
2) Process doing restoration may have different privileges
   than we do (across a sudo, for example, maybe) and
   be unable to remove the memory segment.
   (Due to +s being set on /dev/shm for security reasons)

New strategy is to issue unlink on memory object we still
have an open fd for, so that if either of the above happen
the OS will unlink the object as soon as that fd is closed.

Luckily despite being unable to issue the unlink request,
the process restoring state can always close the fd :).

Especially luckily is that this actually seems to work.  I think :).