libipc: Be sure lock is released if assert fails or we exit.
Introduce ASSERT_WITH_LOCK to do this for us,
replace calls to exit(1) with failing assertions for now.

This is needed to ensure anything done by destructors
that need the lock have a chance of working.

Of course, who knows what the state is since we
are in the process of aborting...