libipc: Lower table size, greatly lower epoll data.  Helps fork().
fork() is slow even when we do nothing from our library
but forward to system's implementation.

I believe this is simply due to our memory footprint,
so this commit attempts to shrink it.

Moving forward, a reasonable simply allocator
would probably be very helpful here, especially
for managing information describing
epoll sets.

(Just because we want to support epoll sets of up
to 1000 doesn't mean we should be statically allocating
a table for that for every epoll file descriptor! :))